refactor: switch convention for Module file names
Module files should not include the word `Module`. So prefer:

`ignition/modules/Lock.ts`

to

`ignition/modules/LockModule.ts`
