+++ packages/hardhat-core/sample-projects/javascript-esm/ignition/modules/Lock.js
