+++ packages/hardhat-core/sample-projects/javascript/ignition/modules/Lock.js
