+++ packages/hardhat-core/sample-projects/typescript-viem/README.md
@@ -9,5 +9,5 @@ npx hardhat help
 npx hardhat test
 REPORT_GAS=true npx hardhat test
 npx hardhat node
-npx hardhat ignition deploy ./ignition/modules/LockModule.ts
+npx hardhat ignition deploy ./ignition/modules/Lock.ts
 ```
