+++ packages/hardhat-core/sample-projects/typescript-viem/ignition/modules/Lock.ts
