+++ packages/hardhat-core/sample-projects/typescript-viem/test/Lock.ts
@@ -6,7 +6,7 @@ import { expect } from "chai";
 import hre from "hardhat";
 import { getAddress, parseGwei } from "viem";
 
-import LockModule from "../ignition/modules/LockModule";
+import LockModule from "../ignition/modules/Lock";
 
 describe("Lock", function () {
   // We define a fixture to reuse the same setup in every test.
