+++ packages/hardhat-core/sample-projects/typescript/ignition/modules/Lock.ts
